+++ hn_talentmatch.client/src/components/ResumePage/ResumePage.jsx
@@ -1,13 +1,13 @@
 import React, { useState } from 'react';
 import ResumeCard from '../resumeCard/ResumeCard';
-import useData from '../../hooks/useData.js';
 import Modal from '../modal/Modal';
 import Pagination from "../pagination/Pagination.jsx";
 import { Container, Title, CardsContainer } from '../../styles/ResumePageStyles';
 import {ActionButton} from "../../styles/CardStyles.js";
+import useResumesData from "../../hooks/useResumesData";
 
 const ResumePage = () => {
-    const { resumeData, loading, error } = useData();
+    const { resumeData, loading, error } = useResumesData();
     const [currentPage, setCurrentPage] = useState(1);
     const [resumesPerPage] = useState(16);
     const [selectedResume, setSelectedResume] = useState(null);
