+++ hn_talentmatch.client/src/hooks/useResumesData.js
@@ -3,7 +3,6 @@ import axios from 'axios';
 
 const useData = () => {
     const [resumeData, setResumeData] = useState([]);
-    const [vacancyData, setVacancyData] = useState([]);
     const [loading, setLoading] = useState(true);
     const [error, setError] = useState(null);
 
@@ -12,9 +11,6 @@ const useData = () => {
             try {
                 const resumeResponse = await axios.get('resumes/GetAll');
                 setResumeData(resumeResponse.data);
-
-                const vacancyResponse = await axios.get('vacancies/GetAll');
-                setVacancyData(vacancyResponse.data);
 
                 setLoading(false);
             } catch (error) {
@@ -27,15 +23,6 @@ const useData = () => {
 
     }, []);
 
-    const getVacancyById = async (id) => {
-        try {
-            const response = await axios.get(`vacancies/GetById?id=${id}`);
-            return response.data;
-        } catch (error) {
-            throw error;
-        }
-    };
-
     const getResumeById = async (id) => {
         try {
             const response = await axios.get(`resumes/GetById?id=${id}`);
@@ -45,7 +32,7 @@ const useData = () => {
         }
     };
 
-    return { resumeData, vacancyData, loading, error, getVacancyById, getResumeById };
+    return { resumeData, loading, error, getResumeById };
 }
 
 export default useData;
